requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.5=pypi_0
aiohttp-cors=0.7.0=pypi_0
aiosignal=1.3.1=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
appnope=0.1.2=py39hecd8cb5_1001
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
bottleneck=1.3.4=py39h67323c0_0
brotli=1.0.9=hb7f2c08_9
brotli-bin=1.0.9=hb7f2c08_9
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.4.26=hecd8cb5_0
cachetools=5.3.1=pypi_0
certifi=2022.6.15=py39hecd8cb5_0
cffi=1.15.1=py39h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=pypi_0
colorful=0.5.5=pypi_0
comm=0.1.2=py39hecd8cb5_0
cryptography=41.0.2=py39h3b477ad_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.7=py39hcec6c5f_0
decorator=5.1.1=pyhd3eb1b0_0
distlib=0.3.7=pypi_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=h0a44026_0
filelock=3.9.0=py39hecd8cb5_0
flwr=1.4.0=pypi_0
fonttools=4.41.1=py39hdc70f33_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.4.0=pypi_0
gettext=0.21.0=he85b6c0_1
giflib=5.2.1=h6c40b1e_3
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py39hd5de756_0
gnutls=3.6.15=hed9c0bf_0
google-api-core=2.11.1=pypi_0
google-auth=2.22.0=pypi_0
googleapis-common-protos=1.60.0=pypi_0
gpustat=1.1=pypi_0
grpcio=1.56.2=pypi_0
icu=58.2=h0a44026_3
idna=3.4=py39hecd8cb5_0
importlib-metadata=6.0.0=py39hecd8cb5_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2023.1.0=ha357a0b_43547
ipykernel=6.19.2=py39h01d92e1_0
ipython=8.12.0=py39hecd8cb5_0
iterators=0.0.2=pypi_0
jedi=0.18.1=py39hecd8cb5_1
jinja2=3.1.2=py39hecd8cb5_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h6c40b1e_1
jsonschema=4.18.4=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter_client=8.1.0=py39hecd8cb5_0
jupyter_core=5.3.0=py39hecd8cb5_0
kiwisolver=1.4.4=py39h92daf61_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hb7f2c08_9
libbrotlidec=1.0.9=hb7f2c08_9
libbrotlienc=1.0.9=hb7f2c08_9
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
libpng=1.6.39=h6c40b1e_0
libsodium=1.0.18=h1de35cc_0
libtasn1=4.19.0=h6c40b1e_0
libtiff=4.5.0=hcec6c5f_2
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxml2=2.10.3=h930c0e2_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.1=py39hca72f7f_0
matplotlib=3.5.3=py39h6e9494a_2
matplotlib-base=3.5.3=py39h35e3ac7_2
matplotlib-inline=0.1.6=py39hecd8cb5_0
mkl=2023.1.0=h59209a4_43558
mkl-service=2.4.0=py39h6c40b1e_1
mkl_fft=1.3.6=py39h07fba90_1
mkl_random=1.2.2=py39h07fba90_1
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.3.0=py39hecd8cb5_0
msgpack=1.0.5=pypi_0
multidict=6.0.4=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=py39hecd8cb5_0
nettle=3.7.3=h230ac6f_1
networkx=3.1=py39hecd8cb5_0
numexpr=2.8.4=py39h47b59a4_1
numpy=1.25.0=py39h827a554_0
numpy-base=1.25.0=py39ha186be2_0
nvidia-ml-py=12.535.77=pypi_0
opencensus=0.11.2=pypi_0
opencensus-context=0.1.3=pypi_0
openh264=2.1.1=h8346a28_0
openssl=3.0.9=hca72f7f_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.4.2=py39he9d5cce_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39hcec6c5f_0
pip=23.2.1=py39hecd8cb5_0
platformdirs=2.5.2=py39hecd8cb5_0
pooch=1.4.0=pyhd3eb1b0_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.36=py39hecd8cb5_0
protobuf=3.20.3=pypi_0
psutil=5.9.0=py39hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-spy=0.3.14=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.12=pypi_0
pygments=2.15.1=py39hecd8cb5_1
pyopenssl=23.2.0=py39hecd8cb5_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.17=h5ee71fb_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=2.0.1=py3.9_0
pytz=2022.1=py39hecd8cb5_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py39hcec6c5f_0
ray=2.6.1=pypi_0
readline=8.2=hca72f7f_0
referencing=0.30.0=pypi_0
requests=2.31.0=py39hecd8cb5_0
rpds-py=0.9.2=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.10.1=py39hf241641_1
setuptools=68.0.0=py39hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
smart-open=6.3.0=pypi_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py39hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h5d9f67b_0
torchaudio=2.0.2=py39_cpu
torchvision=0.15.2=py39_cpu
tornado=6.3.2=py39hdc70f33_0
traitlets=5.7.1=py39hecd8cb5_0
typing_extensions=4.7.1=py39hecd8cb5_0
tzdata=2023c=h04d1e81_0
unicodedata2=15.0.0=py39ha30fb19_0
urllib3=1.26.16=py39hecd8cb5_0
virtualenv=20.21.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py39hecd8cb5_0
xz=5.4.2=h6c40b1e_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h23ab428_0
zipp=3.11.0=py39hecd8cb5_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
